add expenses to Provider
