+++ frontend/src/app/dashboard/page.tsx
@@ -7,5 +7,23 @@ import {useGlobalContext} from "../providers/GlobalProvider"
 export default function DashboardRoute() {
     const {currentExpense} = useGlobalContext()
 
-    return <div>{currentExpense}</div>
+    return (
+        <div className="grid-flow-col-dense gap-10 ">
+            <div
+                className="rounded-md bg-white font-bold flex items-center justify-center flex-1"
+            >{currentExpense}</div>
+            <div
+                className="rounded-md bg-white font-bold flex items-center justify-center flex-1"
+            >{currentExpense}</div>
+            <div
+                className="rounded-md bg-white font-bold flex items-center justify-center flex-1"
+            >{currentExpense}</div>
+            <div
+                className="rounded-md bg-white font-bold flex items-center justify-center flex-1"
+            >{currentExpense}</div>
+            <div
+                className="rounded-md bg-white font-bold flex items-center justify-center flex-1"
+            >{currentExpense}</div>
+        </div>
+    )
 }
