+++ frontend/src/types/types.tsx
@@ -1,0 +1,8 @@
+
+export interface ExpenseItem {
+    name: string
+    amount: number
+    createdAt: string
+    id: string
+    userId: string
+}
